# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Presensi Event Alumni Pesantren API",
        "description": "API documentation for Presensi Event Alumni Pesantren system.",
        "contact": {
            "email": "admin@pesantren.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/presensi/scan": {
            "post": {
                "tags": [
                    "Presensi"
                ],
                "summary": "Scan QR code for attendance",
                "description": "Alumni scans QR code to record attendance. Validates event status, time window, and prevents double scan.",
                "operationId": "scanPresensi",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "qr_token"
                                ],
                                "properties": {
                                    "qr_token": {
                                        "type": "string",
                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Attendance recorded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Presensi berhasil dicatat"
                                        },
                                        "data": {
                                            "properties": {
                                                "presensi": {
                                                    "$ref": "#/components/schemas/Presensi"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Scan failed (inactive, expired, out of window, or double scan)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User not registered for event",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "QR Code tidak dikenali"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "QR token not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/presensi/history": {
            "get": {
                "tags": [
                    "Presensi"
                ],
                "summary": "Get alumni attendance history",
                "description": "Returns the attendance history of the currently logged-in alumni.",
                "operationId": "presensiHistory",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attendance history",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "history": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Presensi"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Ahmad Fauzi"
                    },
                    "gender": {
                        "type": "string",
                        "example": "L"
                    },
                    "status": {
                        "type": "string",
                        "example": "alumni"
                    },
                    "email": {
                        "type": "string",
                        "example": "ahmad@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "081234567890"
                    },
                    "angkatan": {
                        "type": "string",
                        "example": "2015"
                    },
                    "role": {
                        "type": "string",
                        "example": "alumni"
                    },
                    "avatar_url": {
                        "type": "string",
                        "example": "/storage/avatars/avatar123.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error message here"
                    }
                },
                "type": "object"
            },
...
            "Event": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_by": {
                        "type": "integer",
                        "example": 1
                    },
                    "event_title": {
                        "type": "string",
                        "example": "Reuni Akbar 2025"
                    },
                    "description": {
                        "type": "string",
                        "example": "Reuni alumni angkatan 2010-2015"
                    },
                    "location": {
                        "type": "string",
                        "example": "Aula Pesantren"
                    },
                    "event_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-01"
                    },
                    "start_time": {
                        "type": "string",
                        "format": "time",
                        "example": "08:00"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "time",
                        "example": "17:00"
                    },
                    "qr_token": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716"
                    },
                    "qr_code_image": {
                        "type": "string",
                        "example": "qrcodes/550e8400.png"
                    },
                    "qr_code_url": {
                        "type": "string",
                        "example": "http://localhost:8000/storage/qrcodes/550e8400.svg"
                    },
                    "poster_image": {
                        "type": "string",
                        "example": "event-posters/poster123.jpg",
                        "nullable": true
                    },
                    "poster_url": {
                        "type": "string",
                        "example": "http://localhost:8000/storage/event-posters/poster123.jpg",
                        "nullable": true
                    },
                    "status_event": {
                        "type": "string",
                        "example": "active"
                    },
                    "quota": {
                        "type": "integer",
                        "example": 100,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Presensi": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "event_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "scanned_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "event": {
                        "$ref": "#/components/schemas/Event"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}